+++ dsptoolbox/_general_helpers.py
@@ -1341,8 +1341,10 @@ def _get_fractional_impulse_peak_index(
         roots = np.roots(pol)
         # Get only root between 0 and 1
         roots = roots[
-            (roots == roots.real)  # Real roots
-            & (roots <= 1)  # Range
+            # Real roots
+            (roots == roots.real)
+            # Range
+            & (roots <= 1)
             & (roots >= 0)
         ].real
         try:
@@ -1796,32 +1798,34 @@ def _get_correlation_of_latencies(
 
 def __levison_durbin_recursion(
     autocorrelation: NDArray[np.float64],
-) -> tuple[NDArray[np.float64], float]:
-    """Levinson-Durbin recursion to be applied to the autocorrelation
-    estimate.
+) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
+    """Levinson-Durbin recursion to be applied to the autocorrelation estimate.
+    It is always computed along the first (most outer) axis.
 
     Parameters
     ----------
     autocorrelation : NDArray[np.float64]
         Autocorrelation function with only positive lags and length of
         `order + 1`, where `order` corresponds to the order of the AR
-        estimation. It should be a flat array (only one channel).
+        estimation. It can have any shape, but the AR parameters are always
+        computed along the outer axis.
 
     Returns
     -------
     reflection_coefficients : NDArray[np.float64]
-        Denominator coefficients.
-    prediction_error : float
+        Denominator coefficients with shape (coefficient, ...).
+    prediction_error : NDArray[np.float64]
         Variance of the remaining error.
 
     """
-    prediction_error = autocorrelation[0]  # Signal variance
-    autocorr_coefficients = autocorrelation[1:]
-    num_coefficients = len(autocorr_coefficients)
-    ar_parameters = np.zeros(num_coefficients)
+    prediction_error = autocorrelation[0, ...].copy()  # Signal variance
+    autocorr_coefficients = autocorrelation[1:, ...].copy()
+
+    num_coefficients = autocorr_coefficients.shape[0]
+    ar_parameters = np.zeros_like(autocorr_coefficients)
 
     for order in range(num_coefficients):
-        reflection_value = autocorr_coefficients[order]
+        reflection_value = autocorr_coefficients[order].copy()
         if order == 0:
             reflection_coefficient = -reflection_value / prediction_error
         else:
@@ -1831,7 +1835,7 @@ def __levison_durbin_recursion(
                 )
             reflection_coefficient = -reflection_value / prediction_error
         prediction_error *= 1.0 - reflection_coefficient**2.0
-        if prediction_error <= 0:
+        if np.any(prediction_error <= 0):
             raise ValueError("Invalid prediction error: Singular Matrix")
         ar_parameters[order] = reflection_coefficient
 
@@ -1841,7 +1845,7 @@ def __levison_durbin_recursion(
         half_order = (order + 1) // 2
         for lag in range(half_order):
             reverse_lag = order - lag - 1
-            save_value = ar_parameters[lag]
+            save_value = ar_parameters[lag].copy()
             ar_parameters[lag] = (
                 save_value
                 + reflection_coefficient * ar_parameters[reverse_lag]
@@ -1850,8 +1854,76 @@ def __levison_durbin_recursion(
                 ar_parameters[reverse_lag] += (
                     reflection_coefficient * save_value
                 )
+
     # Add first coefficient a0
-    return np.hstack([1.0, ar_parameters]), prediction_error
+    ndim = ar_parameters.ndim
+    pad_width = tuple([(1, 0)] + [(0, 0)] * (ndim - 1))
+    return (
+        np.pad(ar_parameters, pad_width, mode="constant", constant_values=1.0),
+        prediction_error,
+    )
+
+
+def __yw_ar_estimation(
+    time_data: NDArray[np.float64], order: int
+) -> tuple[NDArray[np.float64], NDArray[np.float64]]:
+    """Compute the autoregressive coefficients for an AR process using the
+    Levinson-Durbin recursion to solve the Yule-Walker equations. This is done
+    from the biased autocorrelation.
+
+    Parameters
+    ----------
+    time_data : NDArray[np.float64]
+        Time data with up to three dimensions. The AR parameters are always
+        computed along the first (outer) axis.
+    order : int
+        Recursion order.
+
+    Returns
+    -------
+    NDArray[np.float64]
+        Reflection coefficients with shape (coefficient, ...).
+    NDArray[np.float64]
+        Variance of the remaining error.
+
+    """
+    assert (
+        time_data.ndim <= 3
+    ), "This function only accepts a signal with one, two or three dimensions"
+
+    length_td = time_data.shape[0]
+    if time_data.ndim == 1:
+        autocorrelation = (
+            correlate(time_data, time_data, "full")[
+                length_td - 1 : length_td + order
+            ]
+            / length_td
+        )
+    elif time_data.ndim == 2:
+        autocorrelation = np.zeros((order + 1, time_data.shape[1]))
+        for i in range(time_data.shape[1]):
+            # Biased autocorrelation (only positive lags)
+            autocorrelation[:, i] = (
+                correlate(time_data[:, i], time_data[:, i], "full")[
+                    length_td - 1 : length_td + order
+                ]
+                / length_td
+            )
+    else:
+        autocorrelation = np.zeros(
+            (order + 1, time_data.shape[1], time_data.shape[2])
+        )
+        for ii in range(time_data.shape[2]):
+            for i in range(time_data.shape[1]):
+                # Biased autocorrelation (only positive lags)
+                autocorrelation[:, i, ii] = (
+                    correlate(
+                        time_data[:, i, ii], time_data[:, i, ii], "full"
+                    )[length_td - 1 : length_td + order]
+                    / length_td
+                )
+
+    return __levison_durbin_recursion(autocorrelation)
 
 
 def __burg_ar_estimation(
@@ -1871,7 +1943,8 @@ def __burg_ar_estimation(
     Returns
     -------
     NDArray[np.float64]
-        Denominator (reflection) coefficients.
+        Denominator (reflection) coefficients with shape (coefficient,
+        channel).
     NDArray[np.float64]
         Variances of the prediction error.
 
@@ -1930,4 +2003,4 @@ def __burg_ar_estimation(
         fwd_pred_error = fwd_pred_error[1:]
         bwd_pred_error = bwd_pred_error[:-1]
 
-    return ar_coeffs.squeeze(), den[0]
+    return ar_coeffs.squeeze() if onedim else ar_coeffs, den[0]